PIM-5470: Fix memory leak on rule engine
This version fixes the memory leak directly from the detach function of the objectDetacher object.

Whereas in the previous commit it was defined in a separate bundle and service.

This fix has the advantage of being easily mergable into 1.5 and takes effect everywhere an entity is being detached from doctrine.